renamed a reducer name playing -> player
