+++ src/reducers/index.js
@@ -11,12 +11,12 @@
 import { combineReducers } from 'redux';
 import canvas from '@reducers/canvas.js';
 import palette from '@reducers/palette.js';
-import playing from '@reducers/playing.js';
+import player from '@reducers/player.js';
 
 const reducers = {
   palette,
   canvas,
-  playing
+  player
 };
 const combined = combineReducers(reducers);
 module.exports = combined;
