+++ src/reducers/player.js
